Pluginfy RuboCop RSpec Extra
This PR adds support for RuboCop's Plugin feature.

It replaces the ad-hoc Inject with RuboCop plugins introduced in RuboCop 1.72.

NOTE: Version 1.72.1 or later is specified, as it includes rubocop/rubocop#13837, which removes the Inject from spec_helper.rb. While 1.72.0 is sufficient for runtime, but specifying the patch version 1.72.1 keeps things simple.